+++ pppd/plugins/pppoatm/text2qos.c
@@ -130,7 +130,7 @@ int text2qos(const char *text,struct atm_qos *qos,int flags)
     aal = ATM_NO_AAL;
     do {
 	static const unsigned char aal_number[] = { ATM_AAL0, ATM_AAL5 };
-	int item;
+	size_t item;
 
 	item = fetch(&text,"!none","ubr","cbr","vbr","abr","aal0","aal5",NULL);
 	switch (item) {
